feat: Optimize SIDM isothermal density profile calculations
When seeking the solution for an ISDM isothermal density profile in the presence of baryons, re-use the previous solution for the same node (where available) as the initial guess. Since baryon distributions typically change little from step to step this is usually a good guess and results in more rapid convergence to the solution.